[hw3] refactor page, add page metadata, change favicon
